changed routes to set and get ideas from local storage, also delete them from there

+++ routes/ideas.js
@@ -85,7 +85,11 @@ router.post('/', async (req, res) => {
 
 router.put('/:id', async (req, res) => {
     try {
-        const updatedIdea = await Idea.findByIdAndUpdate(
+        const idea = await Idea.findById(req.params.id);
+
+        // match the usernames
+        if (idea.username === req.body.username) {
+            const updatedIdea = await Idea.findByIdAndUpdate(
             req.params.id,
             {
                 $set: {
@@ -95,8 +99,13 @@ router.put('/:id', async (req, res) => {
             },
             {new: true}
         );
-        res.json({success: true, data: updatedIdea});
-    } catch (error) {
+        return res.json({success: true, data: updatedIdea});
+        }
+
+        // useernames dont match
+        res.status(403).json({success: false, error: 'You are not authorized to edit this resource'});
+        
+        } catch (error) {
         console.log(error);
         res.status(500).json({success: false, error: 'Something went wrong'});
     }
@@ -106,8 +115,15 @@ router.put('/:id', async (req, res) => {
 // delete an idea
 router.delete('/:id', async (req, res) => {
     try {
-        await Idea.findByIdAndDelete(req.params.id);
-    res.json({success: true, data: {}});
+        const idea = await Idea.findById(req.params.id);
+        // match the usernames
+        if (idea.username===req.body.username) {
+            await Idea.findByIdAndDelete(req.params.id);
+            return res.json({success: true, data: {}});
+        }
+        // usernames dont match
+        res.status(403).json({success: false, error: 'You are not authorized to delete this resource'});    
+
 
     } catch (error) {
         console.log(error);
